chore: remove pseudo scope annotations since they don't work
