+++ deployment/src/main/java/com/vaadin/quarkus/deployment/VaadinQuarkusProcessor.java
@@ -41,7 +41,6 @@ import com.vaadin.flow.server.VaadinServlet;
 import com.vaadin.quarkus.QuarkusVaadinServlet;
 import com.vaadin.quarkus.annotation.NormalRouteScoped;
 import com.vaadin.quarkus.annotation.NormalUIScoped;
-import com.vaadin.quarkus.annotation.RouteScoped;
 import com.vaadin.quarkus.annotation.VaadinServiceScoped;
 import com.vaadin.quarkus.annotation.VaadinSessionScoped;
 import com.vaadin.quarkus.context.RouteScopedContext;
@@ -146,11 +145,6 @@ class VaadinQuarkusProcessor {
         return new ContextConfiguratorBuildItem(
                 phase.getContext().configure(NormalRouteScoped.class).normal()
                         .contextClass(RouteScopedContext.class));
-    }
-
-    @BuildStep
-    CustomScopeBuildItem routeScope() {
-        return new CustomScopeBuildItem(RouteScoped.class);
     }
 
     @BuildStep
